move breakpoint methods to behavior
