+++ src/Tribe/Views/V2/View.php
@@ -8,6 +8,7 @@
 
 namespace Tribe\Events\Views\V2;
 
+use Tribe\Events\Views\V2\Behaviors\Breakpoint_Behavior;
 use Tribe\Events\Views\V2\Template\Settings\Advanced_Display;
 use Tribe\Events\Views\V2\Template\Title;
 use Tribe__Container as Container;
@@ -27,6 +28,9 @@ use Tribe__Utils__Array as Arr;
  * @since   4.9.2
  */
 class View implements View_Interface {
+
+	use Breakpoint_Behavior;
+
 	/**
 	 * An instance of the DI container.
 	 *
@@ -1787,108 +1791,6 @@ class View implements View_Interface {
 		$show_datepicker_submit = apply_filters( "tribe_events_views_v2_view_{$this->slug}_show_datepicker_submit", $show_datepicker_submit, $this );
 
 		return $show_datepicker_submit;
-	}
-
-	/**
-	 * Returns the xsmall breakpoint value
-	 *
-	 * @since TBD
-	 *
-	 * @return int
-	 */
-	protected function get_xsmall_breakpoint() {
-		// Default value for xsmall breakpoint is 500px
-		$bp_xsmall = 500;
-
-		/**
-		 * Filters the xsmall breakpoint value.
-		 *
-		 * @since TBD
-		 *
-		 * @param object $bp_xsmall The xsmall breakpoint value.
-		 * @param View   $this      The current View instance being rendered.
-		 */
-		$bp_xsmall = apply_filters( "tribe_events_views_v2_view_breakpoint_xsmall", $bp_xsmall, $this );
-
-		/**
-		 * Filters the xsmall breakpoint value for a specific view.
-		 *
-		 * @since TBD
-		 *
-		 * @param object $bp_xsmall The xsmall breakpoint value.
-		 * @param View   $this      The current View instance being rendered.
-		 */
-		$bp_xsmall = apply_filters( "tribe_events_views_v2_view_{$this->slug}_breakpoint_xsmall", $bp_xsmall, $this );
-
-		return $bp_xsmall;
-	}
-
-	/**
-	 * Returns the medium breakpoint value
-	 *
-	 * @since TBD
-	 *
-	 * @return int
-	 */
-	protected function get_medium_breakpoint() {
-		// Default value for medium breakpoint is 768px
-		$bp_medium = 768;
-
-		/**
-		 * Filters the medium breakpoint value.
-		 *
-		 * @since TBD
-		 *
-		 * @param object $bp_medium The medium breakpoint value.
-		 * @param View   $this      The current View instance being rendered.
-		 */
-		$bp_medium = apply_filters( "tribe_events_views_v2_view_breakpoint_medium", $bp_medium, $this );
-
-		/**
-		 * Filters the medium breakpoint value for a specific view.
-		 *
-		 * @since TBD
-		 *
-		 * @param object $bp_medium The medium breakpoint value.
-		 * @param View   $this      The current View instance being rendered.
-		 */
-		$bp_medium = apply_filters( "tribe_events_views_v2_view_{$this->slug}_breakpoint_medium", $bp_medium, $this );
-
-		return $bp_medium;
-	}
-
-	/**
-	 * Returns the full breakpoint value
-	 *
-	 * @since TBD
-	 *
-	 * @return int
-	 */
-	protected function get_full_breakpoint() {
-		// Default value for full breakpoint is 960px
-		$bp_full = 960;
-
-		/**
-		 * Filters the full breakpoint value.
-		 *
-		 * @since TBD
-		 *
-		 * @param object $bp_full The full breakpoint value.
-		 * @param View   $this    The current View instance being rendered.
-		 */
-		$bp_full = apply_filters( "tribe_events_views_v2_view_breakpoint_full", $bp_full, $this );
-
-		/**
-		 * Filters the full breakpoint value for a specific view.
-		 *
-		 * @since TBD
-		 *
-		 * @param object $bp_full The full breakpoint value.
-		 * @param View   $this    The current View instance being rendered.
-		 */
-		$bp_full = apply_filters( "tribe_events_views_v2_view_{$this->slug}_breakpoint_full", $bp_full, $this );
-
-		return $bp_full;
 	}
 
 	/**
